cleanup & doc
