+++ Cargo.toml
@@ -1,14 +1,14 @@
 [package]
 name = "base85"
-version = "2.0.0"
-authors = ["Jon Yoder <jon@yoder.cloud>"]
+version = "1.0.0"
+authors = ["gerald Hermant"]
 edition = "2021"
-description = "A library for Base85 encoding as described in RFC1924"
+description = "A fast library for Base85 encoding as described in RFC1924"
 readme = "README.md"
-homepage = "https://github.com/darkwyrm/base85"
-repository = "https://github.com/darkwyrm/base85"
+homepage = "https://github.com/geraldhmt/base85"
+repository = "https://github.com/geraldhmt/base85"
 license = "MPL-2.0-no-copyleft-exception"
-keywords = ["base85", "encode", "decode"]
+keywords = ["base85", "rfc1924", "encode", "decode"]
 categories = ["encoding"]
 exclude = [".gitignore"]
 
@@ -17,7 +17,7 @@ thiserror = { version = "2.0.0", default-features = false }
 
 [dev-dependencies]
 criterion = { version = "0.8.2", features = ["html_reports"] }
-rand = "0.8.5"
+rand = "0"
 anyhow = "1.0.71"
 
 [[bench]]
@@ -26,8 +26,4 @@ harness = false
 
 [[bench]]
 name = "encode_noalloc"
-harness = false
-
-[[bench]]
-name = "encode_noalloc_remapped"
 harness = false
